move common IAP components out of pages
