+++ apps/web/src/pages/Home/IAPHomePage/components/Card.tsx
@@ -1,6 +1,6 @@
 import { HTMLAttributes, ReactElement, ReactNode } from "react";
 
-import Heading from "./Heading";
+import Heading from "~/components/IAPHeading/Heading";
 
 interface CardProps {
   title: string;
